requirements.txt
streamlit
google-cloud-speech
google-auth
pyaudio
